requirements.txt
Werkzeug==0.16.0
XlsxWriter==1.2.0
PyMySQL==0.9.3
Flask_Compress==1.4.0
Flask==1.1.1
python-dotenv==0.10.5
